+++ src/test/java/com/braintreegateway/unittest/DisputeTest.java
@@ -25,7 +25,7 @@ public class DisputeTest {
         "<case-number>CASE-12345</case-number>\n " +
         "<created-at type=\"datetime\">2017-06-16T20:44:41Z</created-at>\n " +
         "<currency-iso-code>USD</currency-iso-code>\n " +
-        "<forwarded-comments>Forwarded comments</forwarded-comments>\n" +
+        "<processor-comments>Forwarded comments</processor-comments>\n" +
         "<kind>chargeback</kind>\n " +
         "<merchant-account-id>abc123</merchant-account-id>\n " +
         "<reason>fraud</reason>\n " +
@@ -182,6 +182,7 @@ public class DisputeTest {
         assertEquals(dispute.getCreatedAt().get(Calendar.MINUTE), 44);
         assertEquals(dispute.getCreatedAt().get(Calendar.SECOND), 41);
         assertEquals(dispute.getForwardedComments(), "Forwarded comments");
+        assertEquals(dispute.getProcessorComments(), "Forwarded comments");
         assertEquals(dispute.getMerchantAccountId(), "abc123");
         assertEquals(dispute.getOriginalDisputeId(), "original_dispute_id");
         assertEquals(dispute.getReasonCode(), "83");